refactor(core): delete dead code and simplify template-scaffold
- Remove boilerplate-scanner.ts and boilerplate-types.ts (unused exports)
- Remove exports from index.ts for deleted files
- Inline necessary functions (readBoilerplatesConfig, readBoilerplateConfig)
- Add prompter pass-through to templatizer.process() calls
- Use 'as any' for prompter param until create-gen-app is updated
