+++ pgpm/core/src/index.ts
@@ -10,8 +10,6 @@ export * from './resolution/resolve';
 export * from './workspace/paths';
 export * from './workspace/utils';
 export * from './core/template-scaffold';
-export * from './core/boilerplate-types';
-export * from './core/boilerplate-scanner';
 
 // Export package-files functionality (now integrated into core)
 export * from './files';
